# perf(pkgs-graph): optimize createNode by using Map