This PR rewrites the object spreading of dependencies into using a `Map` as the backing store. Leveraging this excellent test case from @jakebailey https://github.com/jakebailey/DefinitelyTyped/tree/pnpm-workspaces-working the changes in this PR sped up the fresh install time by about 13s on my machine.

| Before | After |
|--:|--:|
| 1:06 min | 0:53min |

Investigation sparked by this twitter thread: https://twitter.com/andhaveaniceday/status/1639851859480551425