@jakebailey 2023-03-26
This PR is missing all of the changes to make it use .get or .has; if it's faster it could probably be just because it's skipping a whole lot of work. Hence, the failing tests.