@marvinhagemeister 2023-03-26
Looks like this PR was a false positive. I accidentally tested it with the `Object.entries()` still being in place and called on a `Map` like @jakebailey outlined. Called on a `Map` that is an empty array so the changes here lead to an incorrect result. Sorry for the confusion, looks like I got excited too early!